requirements.txt
fastapi==0.103.1
pydantic==2.3.0
uvicorn==0.23.2
email-validator==2.0.0.post2
python-jose[cryptography]==3.3.0
python-decouple==3.8
pydantic-settings==2.0.3
python-multipart==0.0.6
passlib==1.7.4
numpy
bcrypt==4.0.1
torch==2.0.1
torchaudio==2.0.2
soundfile==0.12.1
requests
phonemizer==3.2.1
PyYAML==6.0.1
librosa==0.10.1
attridict==0.0.8
pandas==2.0.3
nltk==3.8.1
munch==4.0.0